+++ Assets/Scripts/Card/CardCollection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc19b727cc3b3f040ac90e7c38fb6fe4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
